Refactor move analytics events to the root package
