+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_quiz/domain/analytic/ProblemsLimitReachedModalClickedGoToHomeScreenHyperskillAnalyticEvent.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.step_quiz.domain.analytic.daily_step_completed_modal
+package org.hyperskill.app.step_quiz.domain.analytic
 
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticAction
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticEvent
@@ -7,24 +7,24 @@ import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticRou
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticTarget
 
 /**
- * Represents click on the "Go back" button analytic event.
+ * Represents click on the "Go to homescreen" button in problems limit reached modal analytic event.
  *
  * JSON payload:
  * ```
  * {
  *     "route": "/learn/step/1",
  *     "action": "click",
- *     "part": "daily_step_completed_modal",
- *     "target": "go_back"
+ *     "part": "problems_limit_reached_modal",
+ *     "target": "go_to_home_screen"
  * }
  * ```
  * @see HyperskillAnalyticEvent
  */
-class StepQuizDailyStepCompletedModalClickedGoBackHyperskillAnalyticEvent(
+class ProblemsLimitReachedModalClickedGoToHomeScreenHyperskillAnalyticEvent(
     route: HyperskillAnalyticRoute
 ) : HyperskillAnalyticEvent(
     route,
     HyperskillAnalyticAction.CLICK,
-    HyperskillAnalyticPart.DAILY_STEP_COMPLETED_MODAL,
-    HyperskillAnalyticTarget.GO_BACK
+    HyperskillAnalyticPart.PROBLEMS_LIMIT_REACHED_MODAL,
+    HyperskillAnalyticTarget.GO_TO_HOME_SCREEN
 )
