+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_quiz/domain/analytic/ProblemsLimitReachedModalShownHyperskillAnalyticEvent.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.step_quiz.domain.analytic.problems_limit_reached_modal
+package org.hyperskill.app.step_quiz.domain.analytic
 
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticAction
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticEvent
@@ -7,24 +7,24 @@ import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticRou
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticTarget
 
 /**
- * Represents a hidden analytic event of the problems limit reached modal.
+ * Represents a shown analytic event of the problems limit reached modal.
  *
  * JSON payload:
  * ```
  * {
  *     "route": "/learn/step/1",
- *     "action": "hidden",
- *     "part": "problems_limit_reached_modal",
- *     "target": "close"
+ *     "action": "shown",
+ *     "part": "modal",
+ *     "target": "problems_limit_reached_modal"
  * }
  * ```
  * @see HyperskillAnalyticEvent
  */
-class ProblemsLimitReachedModalHiddenHyperskillAnalyticEvent(
+class ProblemsLimitReachedModalShownHyperskillAnalyticEvent(
     route: HyperskillAnalyticRoute
 ) : HyperskillAnalyticEvent(
     route,
-    HyperskillAnalyticAction.HIDDEN,
-    HyperskillAnalyticPart.PROBLEMS_LIMIT_REACHED_MODAL,
-    HyperskillAnalyticTarget.CLOSE
+    HyperskillAnalyticAction.SHOWN,
+    HyperskillAnalyticPart.MODAL,
+    HyperskillAnalyticTarget.PROBLEMS_LIMIT_REACHED_MODAL
 )
