+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_quiz/domain/analytic/StepQuizDailyStepCompletedModalShownHyperskillAnalyticEvent.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.step_quiz.domain.analytic.problems_limit_reached_modal
+package org.hyperskill.app.step_quiz.domain.analytic
 
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticAction
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticEvent
@@ -7,24 +7,24 @@ import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticRou
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticTarget
 
 /**
- * Represents click on the "Go to homescreen" button in problems limit reached modal analytic event.
+ * Represents a shown analytic event of the daily step completed modal.
  *
  * JSON payload:
  * ```
  * {
  *     "route": "/learn/step/1",
- *     "action": "click",
- *     "part": "problems_limit_reached_modal",
- *     "target": "go_to_home_screen"
+ *     "action": "shown",
+ *     "part": "modal",
+ *     "target": "daily_step_completed_modal"
  * }
  * ```
  * @see HyperskillAnalyticEvent
  */
-class ProblemsLimitReachedModalClickedGoToHomeScreenHyperskillAnalyticEvent(
+class StepQuizDailyStepCompletedModalShownHyperskillAnalyticEvent(
     route: HyperskillAnalyticRoute
 ) : HyperskillAnalyticEvent(
     route,
-    HyperskillAnalyticAction.CLICK,
-    HyperskillAnalyticPart.PROBLEMS_LIMIT_REACHED_MODAL,
-    HyperskillAnalyticTarget.GO_TO_HOME_SCREEN
+    HyperskillAnalyticAction.SHOWN,
+    HyperskillAnalyticPart.MODAL,
+    HyperskillAnalyticTarget.DAILY_STEP_COMPLETED_MODAL
 )
